[Patchset 4] UI: seamlessly migrate to catapult UI for legacy traces
When a legacy trace is detected replaces the body with the old UI.

Bug: 117977226

+++ tools/trace_to_text/ftrace_event_formatter.cc
@@ -3396,7 +3396,7 @@ std::string FormatPrefix(uint64_t timestamp,
 
 std::string FormatFtraceEvent(
     uint64_t timestamp,
-    size_t cpu,
+    uint32_t cpu,
     const protos::FtraceEvent& event,
     const std::unordered_map<uint32_t /*tid*/, uint32_t /*tgid*/>& thread_map) {
   // Sched_switch events contain the thread name so use that in the prefix.
